requirements.txt
Flask>=2.0
google-genai
python-dotenv
